requirements.txt
paho-mqtt
websockets
pymongo